+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Podemos ter pequenas breaking changes sem alterar o `major` version, apesar de s
 
 ## Não publicado
 ### Adicionado
-`QasUploader`: adicionado método de `reset` caso queira resetar o campo.
+`QasUploader`: adicionado método de `reset` caso queira resetar o campo externamente.
 
 ### Corrigido
 `QasUploader`: corrigido valor do emit ao remover um item do uploader, pois estava quebrando no caso do uso `use-object-model`.